Add support for blueprints as subdirectories
Other than linking blueprints as submodules, this commit adds support
for blueprints as subdirectories in the main repo.

A validation will be triggered by each git change except changes from
the anthoscli-release-candidates.json file. It gets the updated
file list from the commit and gets the corresponding packages for
validation. In the end, it updated the release candidate tag.